reduce ts deps - back out index for query-properties - move annotations out of malloy-element
