+++ packages/malloy/src/lang/ast/index.ts
@@ -82,7 +82,25 @@ export * from './query-elements/existing-query';
 export * from './query-elements/full-query';
 export * from './query-items/field-declaration';
 export * from './query-items/field-references';
-export * from './query-properties/index-query-properties';
+export * from './query-properties/aggregate';
+export * from './query-properties/calculate';
+export * from './query-properties/declare-fields';
+export * from './query-properties/dimensions';
+export * from './query-properties/extend';
+export * from './query-properties/filters';
+export * from './query-properties/group-by';
+export * from './query-properties/indexing';
+export * from './query-properties/joins';
+export * from './query-properties/limit';
+export * from './query-properties/measures';
+export * from './query-properties/nest-reference';
+export * from './query-properties/nest';
+export * from './query-properties/nests';
+export * from './query-properties/ordering';
+export * from './query-properties/project-statement';
+export * from './query-properties/qop-desc';
+export * from './query-properties/sampling';
+export * from './query-properties/top';
 export * from './sources/named-source';
 export * from './sources/named-source-or-query';
 export * from './sources/query-source';
@@ -91,6 +109,7 @@ export * from './sources/sql-source';
 export * from './sources/table-source';
 export * from './sql-elements/sql-statement';
 export * from './sql-elements/sql-string';
+export * from './types/annotation-elements';
 export * from './types/comparison';
 export * from './types/source-desc';
 export * from './types/source-property';
