--- packages/malloy/src/lang/ast/query-properties/index-query-properties.ts
@@ -1,20 +1,0 @@
-export * from './aggregate';
-export * from './calculate';
-export * from './declare-fields';
-export * from './dimensions';
-export * from './extend';
-export * from './filters';
-export * from './group-by';
-export * from './indexing';
-export * from './joins';
-export * from './limit';
-export * from './measures';
-export * from './nest';
-export * from './nest-reference';
-export * from './nests';
-export * from './nests';
-export * from './ordering';
-export * from './project-statement';
-export * from './qop-desc';
-export * from './sampling';
-export * from './top';
