+++ packages/malloy/src/lang/ast/types/source-property.ts
@@ -28,8 +28,9 @@ import {FieldListEdit} from '../source-properties/field-list-edit';
 import {Renames} from '../source-properties/renames';
 import {PrimaryKey} from '../source-properties/primary-key';
 import {Turtles} from '../source-properties/turtles';
-import {MalloyElement, ObjectAnnotation} from './malloy-element';
+import {MalloyElement} from './malloy-element';
 import {TimezoneStatement} from '../source-properties/timezone-statement';
+import {ObjectAnnotation} from './annotation-elements';
 
 export type SourceProperty =
   | Filter
